feat(@cubejs-backend/api-gateway): Migrate to TS initial
