+++ packages/cubejs-api-gateway/.gitignore
@@ -1,0 +1,1 @@
+dist
