+++ packages/cubejs-api-gateway/package.json
@@ -11,14 +11,20 @@
   "engines": {
     "node": ">=8.11.1"
   },
+  "main": "dist/src/index.js",
+  "typings": "dist/src/index.d.ts",
   "scripts": {
     "test": "npm run unit",
     "unit": "jest",
-    "lint": "eslint *.js",
-    "lint:fix": "eslint --fix *.js"
+    "tsc": "tsc",
+    "watch": "tsc -w",
+    "lint": "linter src/* test/* --ext .ts,.js",
+    "lint:fix": "linter --fix src/* test/* --ext .ts,.js"
   },
-  "main": "index.js",
-  "types": "index.d.ts",
+  "files": [
+    "README.md",
+    "dist/src/*"
+  ],
   "dependencies": {
     "@hapi/joi": "^15.1.1",
     "body-parser": "^1.19.0",
@@ -31,13 +37,22 @@
     "uuid": "^3.3.3"
   },
   "devDependencies": {
-    "eslint": "^6.8.0",
-    "eslint-plugin-node": "^5.2.1",
+    "@cubejs-backend/linter": "^0.24.2",
+    "@types/express": "^4.17.9",
+    "@types/hapi__joi": "^15.0.4",
+    "@types/jest": "^25.2.3",
+    "@types/jsonwebtoken": "^8.5.0",
+    "@types/moment": "^2.13.0",
+    "@types/ramda": "^0.27.32",
+    "@types/supertest": "^2.0.10",
     "express": "^4.17.1",
-    "jest": "^24.9.0",
-    "mocha": "^8.0.1",
+    "jest": "^25.5.4",
     "should": "^13.2.3",
-    "supertest": "^4.0.2"
+    "supertest": "^4.0.2",
+    "typescript": "^4.1.2"
   },
-  "license": "Apache-2.0"
+  "license": "Apache-2.0",
+  "eslintConfig": {
+    "extends": "../cubejs-linter"
+  }
 }
